refactor : CueStatusView 리팩토링함.

+++ Audio2Tag/old/Cue/Model/SplitMusicModel.swift
@@ -11,5 +11,5 @@ import Foundation
 struct SplitMusicModel : Identifiable {
     let id = UUID()
     let name: String
-    var status: Int
+    var status: Double
 }
